Moved File and Audio to block-components folder
