+++ packages/react-notion-x/src/context.tsx
@@ -3,12 +3,14 @@ import * as React from 'react'
 import { ExtendedRecordMap } from 'notion-types'
 
 import { Alias } from './block-components/alias'
+import { Audio } from './block-components/audio'
 import { Bookmark } from './block-components/bookmark'
 import { Callout } from './block-components/callout'
 import { Column } from './block-components/column'
 import { ColumnList } from './block-components/column-list'
 import { Divider } from './block-components/divider'
 import { Drive } from './block-components/drive'
+import { File } from './block-components/file'
 import { TextHeader } from './block-components/header'
 import { List } from './block-components/list'
 import { Page } from './block-components/page'
@@ -20,10 +22,8 @@ import { TextBlock } from './block-components/text'
 import { ToDo } from './block-components/to-do'
 import { Toggle } from './block-components/toggle'
 import { AssetWrapper } from './components/asset-wrapper'
-import { Audio } from './components/audio'
 import { Checkbox as DefaultCheckbox } from './components/checkbox'
 import { EOI } from './components/eoi'
-import { File } from './components/file'
 import { Header } from './components/header'
 import { wrapNextImage, wrapNextLink } from './next'
 import {
